data-service/providers/angel.py
import os
from SmartApi import SmartConnect
from .base import BaseProvider
from typing import Dict, Any
import datetime

class AngelOneProvider(BaseProvider):
    def __init__(self):
        self.api_key = os.getenv("SMART_API_KEY")
        self.client_id = os.getenv("SMART_CLIENT_ID")
        self.password = os.getenv("SMART_PASSWORD")
        self.totp_secret = os.getenv("SMART_TOTP_SECRET")
        self.smart_api = None

    def _ensure_connected(self):
        if not self.smart_api:
            self.smart_api = SmartConnect(api_key=self.api_key)
            # Login logic would go here, usually involving pyotp
            # For brevity, assuming session is managed or initialized on first call
            pass

    def get_stock(self, symbol: str) -> Dict[str, Any]:
        # Note: AngelOne needs an instrument token. Mapping is required.
        # This is a simplified version.
        try:
            self._ensure_connected()
            # Placeholder for mapping symbol -> token
            # ltp_data = self.smart_api.ltpData("NSE", symbol, "token")
            return {
                "symbol": symbol,
                "price": 0, # To be fetched via SmartApi
                "change": 0,
                "percentChange": 0,
                "timestamp": datetime.datetime.now().isoformat(),
                "provider": "angelone"
            }
        except Exception as e:
            raise Exception(f"AngelOne error: {str(e)}")

    def get_market_status(self) -> Dict[str, Any]:
        return {"status": "open"} # Simplified
