data-service/requirements.txt
fastapi
uvicorn
jugaad-data
requests
python-dotenv
smartapi-python
